# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "InsightExtract API",
        "description": "API for PDF-to-flashcard conversion with spaced repetition",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/cards/{card_id}": {
            "patch": {
                "tags": [
                    "cards"
                ],
                "summary": "Update Card",
                "description": "Update card content.",
                "operationId": "update_card_api_cards__card_id__patch",
                "parameters": [
                    {
                        "name": "card_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Card Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CardUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "cards"
                ],
                "summary": "Delete Card",
                "description": "Delete a card and its associated reviews.",
                "operationId": "delete_card_api_cards__card_id__delete",
                "parameters": [
                    {
                        "name": "card_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Card Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cards/{card_id}/review": {
            "post": {
                "tags": [
                    "cards"
                ],
                "summary": "Review Card",
                "description": "Record a card review and update SRS parameters.",
                "operationId": "review_card_api_cards__card_id__review_post",
                "parameters": [
                    {
                        "name": "card_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Card Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CardReview"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CardResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "deck_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Deck Id"
                    },
                    "type": {
                        "$ref": "#/components/schemas/FlashcardType"
                    },
                    "front": {
                        "type": "string",
                        "title": "Front"
                    },
                    "back": {
                        "type": "string",
                        "title": "Back"
                    },
                    "citations": {
                        "items": {
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Citations"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags"
                    },
                    "ease": {
                        "type": "number",
                        "title": "Ease"
                    },
                    "interval": {
                        "type": "integer",
                        "title": "Interval"
                    },
                    "repetition": {
                        "type": "integer",
                        "title": "Repetition"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Due Date"
                    },
                    "last_reviewed_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Reviewed At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "deck_id",
                    "type",
                    "front",
                    "back",
                    "citations",
                    "tags",
                    "ease",
                    "interval",
                    "repetition",
                    "due_date"
                ],
                "title": "CardResponse",
                "description": "Card response schema."
            },
            "CardReview": {
                "properties": {
                    "quality": {
                        "type": "integer",
                        "maximum": 5.0,
                        "minimum": 0.0,
                        "title": "Quality"
                    }
                },
                "type": "object",
                "required": [
                    "quality"
                ],
                "title": "CardReview",
                "description": "Card review schema."
            },
            "CardUpdate": {
                "properties": {
                    "front": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Front"
                    },
                    "back": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Back"
                    },
                    "tags": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tags"
                    }
                },
                "type": "object",
                "title": "CardUpdate",
                "description": "Card update schema."
            },
...
            "FlashcardType": {
                "type": "string",
                "enum": [
                    "qa",
                    "cloze",
                    "truefalse"
                ],
                "title": "FlashcardType",
                "description": "Flashcard type enum."
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}